+++ htmlelements-samples/src/main/java/ru/yandex/qatools/htmlelements/samples/RefreshMatcherUsage.java
@@ -9,9 +9,9 @@ import ru.yandex.qatools.htmlelements.samples.pages.MainPage;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.not;
-import static ru.yandex.qatools.htmlelements.matchers.RefreshMatcherDecorator.withPrerefresh;
-import static ru.yandex.qatools.htmlelements.matchers.WaitForMatcherDecorator.withWaitFor;
 import static ru.yandex.qatools.htmlelements.matchers.WebElementMatchers.hasText;
+import static ru.yandex.qatools.htmlelements.matchers.decorators.RefreshMatcherDecorator.withPrerefresh;
+import static ru.yandex.qatools.htmlelements.matchers.decorators.WaitForMatcherDecorator.withWaitFor;
 
 /**
  * Created with IntelliJ IDEA.
